requirements.txt
aiohttp 
aioredis
fastapi
fastapi-cache2
bs4
uvicorn
pydantic-settings
httpx
